requirements.txt
pyvmomi
pymongo
pyyaml
fastapi
uvicorn
redis[async]
httpx